split components into smaller pieces
